params interface restriction and query format functions
